Update Section 4.4 TVM Security Requirements to address threat model
There are two updates here to the second line in the tabel labeled Memory
Confidentiality. The first is to change MMU to xMMU because the IO/MMU may
also be affected. The second is to make this optional. The line above it
and this line should probably be combined. The previous line subsummes
this line. IF you prevent non-TCB components from reading TEE memory it
does not matter if the TEE memory is encrypted.